Harden template and regex fallback scanning
